+++ app/javascript/vue/components/radials/annotator/components/depiction_annotator.vue
@@ -194,6 +194,7 @@ import OtuPicker from 'components/otu/otu_picker/otu_picker'
 import RadialAnnotator from 'components/radials/annotator/annotator.vue'
 import FilterImage from 'tasks/images/filter/components/filter'
 import SmartSelector from 'components/ui/SmartSelector'
+import { Depiction } from 'routes/endpoints'
 
 export default {
   mixins: [CRUD, annotatorExtend],
@@ -245,6 +246,11 @@ export default {
           value: 'TaxonName',
           label: 'Taxon name',
           url: '/taxon_names/autocomplete'
+        },
+        {
+          value: 'Person',
+          label: 'Person',
+          url: '/people/autocomplete'
         }
       ],
       isDataDepiction: false,
@@ -274,15 +280,18 @@ export default {
         this.depiction.depiction_object_type = this.selectedType.value
         this.depiction.depiction_object_id = this.selectedObject.id
       }
-      this.update(`/depictions/${this.depiction.id}`, { depiction: this.depiction }).then(response => {
+
+      Depiction.update(this.depiction.id, { depiction: this.depiction }).then(response => {
         const index = this.list.findIndex(element => this.depiction.id === element.id)
 
         if (this.updateObjectType) {
-          delete this.list[index]
+          this.list.splice(index, 1)
         } else {
           this.list[index] = response.body
         }
         this.depiction = undefined
+
+        TW.workbench.alert.create('Depiction was successfully updated.', 'notice')
       })
     },
 
@@ -299,7 +308,7 @@ export default {
         is_metadata_depiction: this.isDataDepiction
       }
 
-      this.create('/depictions.json', { depiction: depiction }).then(({ body }) => {
+      Depiction.create({ depiction }).then(({ body }) => {
         this.list.push(body)
         TW.workbench.alert.create('Depiction was successfully created.', 'notice')
       })